# src/docs/swagger.yaml
swagger: "2.0"
info:
  description: "This is the documentation for the NIMPI project, search microservice"
  version: "1.0.0"
  title: "NIMPI-search"
basePath: "/api"

tags:
- name: "Search"
  description: "Document search routes"
- name: "Operational"
  description: "Routes to operational functions in the system"

schemes:
- "https"
- "http"

paths:
  /indexing:
    post:
      tags:
      - "Operational"
      summary: "Index documents"
      description: "Index documents from mongoDB"
      operationId: "indexing"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        '200':
          description: OK
        '408':
          description: Sorry unable to connect to MongoDB Error
        '503':
          description: Internal Error Server.
      security:
        - basicAuth: []
     
  /v1/document/findbyterm:
    get:
      tags:
      - "Search"
      summary: "Finds document by term"
      description: ""
      operationId: "findDocumentsByTerm"
      produces:
      - "application/json"
      parameters:
        - in: query
          name: term
          required: true
          type: integer
          format: text
          description: Find by free terms
        - in: query
          name: size
          required: true
          type: string
          format: number
          description: page size
        - in: query
          name: from
          required: true
          type: integer
          format: number
          description: From document
      responses:
        '200':
          description: "successful operation"
        '401':
          description: Invalid credentials. Please consult the administrator.
        '400':
          description: Error fetching documents in Elasticsearch
        '500':
          description: Internal Error Server.
  
  /v1/document/findbymetadata:
    get:
      tags:
      - "Search"
      summary: "Finds document by metadata"
      description: "Finds document by metadata"
      operationId: "findDocumentsByDescripition"
      produces:
      - "application/json"
      parameters:
        - in: query
          name: size
          type: integer
          description: Size page
        - in: query
          name: from
          type: integer
          description: From Document
        - in: query
          name: dateInitial
          type: string
          description: Find by date initial, YYYY-mm-dd
        - in: query
          name: dateFinal
          type: string
          description: Find by date final, YYYY-mm-dd
        - in: query
          name: publisherName
          type: string
          description: Find by publisher, data in text format
        - in: query
          name: type
          type: string
          description: Find by type
        - in: query
          name: tagg
          type: string
          description: Find by tags
        - in: query
          name: title
          type: string
          description: Find by title
        - in: query
          name: description
          type: string
          description: Find by description
      responses:
        '200':
          description: "successful operation"
        '401':
          description: Invalid credentials. Please consult the administrator.
        '502':
          description: Bad Gateway.
        '500':
          description: Internal Error Server.
  /v1/autocomplete:
    get:
      tags:
      - "Search"
      summary: "Autocomplete find by title"
      description: "Autocomplete to title"
      operationId: "autocomplete"
      produces:
      - "application/json"
      parameters:
        - in: query
          name: text
          type: string
          description: text to complete
      responses:
        '200':
          description: "successful operation"
        '401':
          description: Invalid credentials. Please consult the administrator.
        '502':
          description: Bad Gateway.
        '500':
          description: Internal Error Server.

  /v1/didyoumean:
    get:
      tags:
      - "Search"
      summary: "Did you mean?"
      description: "Did you mean?"
      operationId: "didyoumean"
      produces:
      - "application/json"
      parameters:
        - in: query
          name: text
          type: string
          description: text to did you mean
      responses:
        '200':
          description: "successful operation"
        '401':
          description: Invalid credentials. Please consult the administrator.
        '502':
          description: Bad Gateway.
        '500':
          description: Internal Error Server.

securityDefinitions:
  basicAuth:
    type: basic
# To apply Basic auth to the whole API:
security:
  - basicAuth: []
  
  
definitions:
  Document:
    type: "object"
    properties:
      id:
        type: "integer"
      title:
        type: "string"
        xml:
          name: "tag"
          wrapped: true
      description:
        type: "string"
      date:
        type: "string"
        format: "date-time"
      year:
        type: "integer"
      publisherId:
        type: "integer"
      type:
        type: "string"
      path:
        type: "string"
      dateCreated:
        type: "string"
        format: "date-time"
      lastModification:
        type: "string"
        format: "date-time"
      tags:
        type: "string"
      userId:
        type: "integer"
      articles:
        type: "array"
        items:
          $ref: '#/definitions/Articles'
        
  
  Publisher:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        
  Articles:
    type: "object"
    properties:
      id:
        type: "integer"
      title:
        type: "string"
        xml:
          name: "tag"
          wrapped: true
      subtitle:
        type: "string"
      authors:
        type: "array"
        items:
          $ref: "#/definitions/Publisher"
      content:
        type: "string"
      imagesList:
        type: "string"
      documentId:
        type: "integer"
  
externalDocs:
  description: "Find out more about Swagger"
  url: "http://nimpi.ifro.edu.br/"
  